backend-ai-detector/requirements.txt
fastapi
uvicorn
google-cloud-aiplatform
python-multipart
scikit-learn
pandas
numpy
joblib